Fix issue blocking parallel execution
The issue caused by fact that System.Web.Hosting.HostingEnvironment
can't be initialized twice. So lock is added to FakeHostEnvironment
preventing parallel instantiation. Also now FakeHostEnvironment will
throw if instantiated twice in the same thread because of lock nesting.
Yes, this fix doesn't not allow true parallel execution, but with it
nobody should do anyting to run the tests with enable parallel
execution. The other option is to not instantiate
System.Web.Hosting.HostingEnvironment if it is already initialized but
in this case FakeHttpContext instances which sets base path will affect
each others, so the only solutions is ask everybody disable parallel
execution or disable it by locking.

Fix for #9